Adding PCB shape generator
Adding first iteration of the PCB 3D shape generator. The generator cuts the part of the PCB board that has two quarter-holes and the space in-between.
The config command from gen is not utilised yet. The center of circles(cylinders) should be on a grid poins, with circle's radiuses being an integer number in grid step units.