Adding 2D strip geometry
Adding generator to make 2D strip geometry for the Ew calculation. The generator draws 2 planes coll/ind. There are two supported patterns : 1) with hole in the middle of the strip 2) holes are on each side of the strip

conf parameter in generator is not yet utilised. Thus, all of parameters are hardcoded and code works with only provided domain